Add user Permissions and cart components

+++ backend/src/resolvers/Mutation.js
@@ -2,11 +2,10 @@ const bcrypt = require("bcryptjs");
 const jwt = require("jsonwebtoken");
 const { randomBytes } = require("crypto");
 const { promisify } = require("util");
-const { createJWT } = require("../utils");
 const { transport, makeEmail } = require("../mail");
+const { hasPermission } = require("../utils");
 const Mutation = {
   async createItem(parent, args, ctx, info) {
-    //TODO check authentication before writing to db
     if (!ctx.request.userId) {
       throw new Error("You must be logged in to do that");
     }
@@ -38,9 +37,16 @@ const Mutation = {
   },
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
-    const item = await ctx.db.query.item({ where }, `{id title}`);
+    const item = await ctx.db.query.item({ where }, `{id title user {id}}`);
 
-    //TODO check permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ["ADMIN", "ITEMDELETE"].includes(permission)
+    );
+
+    if (!ownsItem || !hasPermissions) {
+      throw new Error("I can't let you do that");
+    }
 
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -139,6 +145,75 @@ const Mutation = {
       maxAge: 1000 * 60 * 60 * 24 * 365
     });
     return updatedUser;
+  },
+  async updatePermissions(parent, args, ctx, info) {
+    if (!ctx.request.userId) {
+      throw new Error("You must be logged in");
+    }
+    const currentUser = await ctx.db.query.user(
+      {
+        where: { id: ctx.request.userId }
+      },
+      info
+    );
+    hasPermission(currentUser, ["ADMIN", "PERMISSIONUPDATE"]);
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: { set: args.permissions }
+        },
+        where: {
+          id: args.userId
+        }
+      },
+      info
+    );
+  },
+  async addToCart(parent, args, ctx, info) {
+    const { userId } = ctx.request;
+    if (!userId) {
+      throw new Error("You must be logged in");
+    }
+    const [currentCartItem] = await ctx.db.query.cartItems(
+      { where: { user: { id: userId }, item: { id: args.id } } },
+      info
+    );
+    if (!currentCartItem) {
+      return ctx.db.mutation.createCartItem(
+        {
+          data: {
+            user: { connect: { id: userId } },
+            item: { connect: { id: args.id } },
+            quantity: 1
+          }
+        },
+        info
+      );
+    }
+    return ctx.db.mutation.updateCartItem(
+      {
+        where: { id: currentCartItem.id },
+        data: {
+          quantity: currentCartItem.quantity + 1
+        }
+      },
+      info
+    );
+  },
+  async removeFromCart(parent, args, ctx, info) {
+    const { userId } = ctx.request;
+    if (!userId) {
+      throw new Error("You must be logged in");
+    }
+    const cartItem = await ctx.db.query.cartItem(
+      { where: { id: args.id } },
+      `{id, user {id}}`
+    );
+    if (!cartItem) throw new Error("No Item Found");
+    if (cartItem.user.id !== userId) {
+      throw new Error("This cart item does not belong to you");
+    }
+    return ctx.db.mutation.deleteCartItem({ where: { id: cartItem.id } }, info);
   }
 };
